refactor: split firebird.c into focused compilation units (closes #57)
Split the monolithic 3668-line firebird.c into 5 focused files:
- fbird_error.c (218 lines): error/utility PHP functions + internal helpers
- fbird_connection.c (604 lines): connection management
- fbird_transaction.c (1078 lines): transaction management
- fbird_batch.c (684 lines): batch API (FB 4.0+)
- firebird.c (1201 lines): module core, arginfo, function table, MINIT/MINFO

New headers:
- php_fbird_connection.h: destructor callbacks for le_link/le_plink
- php_fbird_transaction.h: destructor callback for le_trans
- php_fbird_batch.h: destructor callback for le_batch (FB 4.0+)

config.m4: add 4 new .c files to PHP_NEW_EXTENSION source list

Aligns with existing per-concern split pattern (fbird_service.c, fbird_blobs.c, etc.)
Zero public API changes. All function signatures unchanged.
