+++ firebird.c
@@ -23,6 +23,9 @@
 #include <math.h>
 #include "firebird_utils.h"
 #include "fbird_datetime.h"
+#include "php_fbird_connection.h"
+#include "php_fbird_transaction.h"
+#include "php_fbird_batch.h"
 
 #define ROLLBACK    0
 #define COMMIT      1
@@ -585,518 +588,7 @@ const char *_fbird_res_type_name(int type) {
 }
 
 
-PHP_FUNCTION(fbird_errmsg)
-{
-	if (zend_parse_parameters_none() == FAILURE) {
-		return;
-	}
-
-	if (IBG(sql_code) != 0) {
-		RETURN_STRING(IBG(errmsg));
-	}
-
-	RETURN_FALSE;
-}
-
-PHP_FUNCTION(fbird_get_client_version)
-{
-	RETURN_DOUBLE((double)IBG(client_major_version) + (double)IBG(client_minor_version) / 10);
-}
-
-PHP_FUNCTION(fbird_get_client_major_version)
-{
-	RETURN_LONG(IBG(client_major_version));
-}
-
-PHP_FUNCTION(fbird_get_client_minor_version)
-{
-	RETURN_LONG(IBG(client_minor_version));
-}
-
-PHP_FUNCTION(fbird_errcode)
-{
-	if (zend_parse_parameters_none() == FAILURE) {
-		return;
-	}
-
-	if (IBG(sql_code) != 0) {
-		RETURN_LONG(IBG(sql_code));
-	}
-	RETURN_FALSE;
-}
-
-PHP_FUNCTION(fbird_sqlstate)
-{
-	char sqlstate[6]; /* 5 chars + null terminator */
-
-	if (zend_parse_parameters_none() == FAILURE) {
-		return;
-	}
-
-	/* Check if there is an error to report */
-	if (IBG(sql_code) == 0) {
-		RETURN_FALSE;
-	}
-
-	/* Call fb_sqlstate to get the SQLSTATE code from the status vector */
-	fb_sqlstate(sqlstate, IB_STATUS);
-
-	/* fb_sqlstate always returns a 5-character string, with "00000" for success */
-	if (sqlstate[0] == '0' && sqlstate[1] == '0' && sqlstate[2] == '0' &&
-	    sqlstate[3] == '0' && sqlstate[4] == '0') {
-		/* No error state */
-		RETURN_FALSE;
-	}
-
-	RETURN_STRINGL(sqlstate, 5);
-}
-
-PHP_FUNCTION(fbird_escape_string)
-{
-	zend_string *str;
-
-	ZEND_PARSE_PARAMETERS_START(1, 1)
-		Z_PARAM_STR(str)
-	ZEND_PARSE_PARAMETERS_END();
-
-	/* Count single quotes to calculate required buffer size */
-	size_t quote_count = 0;
-	const char *p = ZSTR_VAL(str);
-	size_t len = ZSTR_LEN(str);
-
-	for (size_t i = 0; i < len; i++) {
-		if (p[i] == '\'') {
-			quote_count++;
-		}
-	}
-
-	/* If no quotes, return copy of original string */
-	if (quote_count == 0) {
-		RETURN_STR_COPY(str);
-	}
-
-	/* Allocate buffer for escaped string (original + extra quotes) */
-	zend_string *escaped = zend_string_alloc(len + quote_count, 0);
-	char *out = ZSTR_VAL(escaped);
-
-	/* Copy string, doubling single quotes */
-	for (size_t i = 0; i < len; i++) {
-		if (p[i] == '\'') {
-			*out++ = '\'';
-			*out++ = '\'';
-		} else {
-			*out++ = p[i];
-		}
-	}
-	*out = '\0';
-
-	RETURN_NEW_STR(escaped);
-}
-
-PHP_FUNCTION(fbird_set_exception_mode)
-{
-	zend_long mode;
-
-	if (zend_parse_parameters(ZEND_NUM_ARGS(), "l", &mode) == FAILURE) {
-		return;
-	}
-
-	if (mode != FBIRD_EXCEPTION_MODE_SILENT && mode != FBIRD_EXCEPTION_MODE_THROW) {
-		php_error_docref(NULL, E_WARNING,
-			"Invalid mode (use FBIRD_EXCEPTION_MODE_SILENT or FBIRD_EXCEPTION_MODE_THROW)");
-		RETURN_FALSE;
-	}
-
-	IBG(exception_mode) = (int)mode;
-	RETURN_TRUE;
-}
-
-PHP_FUNCTION(fbird_get_exception_mode)
-{
-	if (zend_parse_parameters_none() == FAILURE) {
-		return;
-	}
-
-	RETURN_LONG(IBG(exception_mode));
-}
-
-PHP_METHOD(FirebirdException, getSqlState)
-{
-	char sqlstate[6]; /* 5 chars + null terminator */
-
-	if (zend_parse_parameters_none() == FAILURE) {
-		return;
-	}
-
-	/* Call fb_sqlstate to get the SQLSTATE code from the status vector */
-	fb_sqlstate(sqlstate, IB_STATUS);
-
-	/* Always return the SQLSTATE (even if "00000" for compatibility) */
-	RETURN_STRINGL(sqlstate, 5);
-}
-
-static const zend_function_entry firebird_exception_methods[] = {
-	PHP_ME(FirebirdException, getSqlState, arginfo_firebird_exception_getSqlState, ZEND_ACC_PUBLIC)
-	PHP_FE_END
-};
-
-/* print firebird error and save it for fbird_errmsg() */
-void _php_fbird_error(void)
-{
-	char *s = IBG(errmsg);
-	const ISC_STATUS *statusp = IB_STATUS;
-	size_t msg_len;
-
-	IBG(sql_code) = isc_sqlcode(IB_STATUS);
-
-	msg_len = strlen(IBG(errmsg));
-	while (msg_len < MAX_ERRMSG && fb_interpret(s, MAX_ERRMSG - msg_len - 1, &statusp)) {
-		msg_len = strlen(s);
-		s[msg_len] = ' ';
-		s[msg_len + 1] = '\0';
-		msg_len = s - IBG(errmsg) + msg_len + 1;
-		s = IBG(errmsg) + msg_len;
-	}
-
-	/* Check runtime exception_mode first, fallback to INI setting */
-	if (IBG(exception_mode) == FBIRD_EXCEPTION_MODE_THROW || INI_BOOL("fbird.enable_exceptions")) {
-		zend_throw_exception(firebird_exception_ce, IBG(errmsg), IBG(sql_code));
-	} else {
-		php_error_docref(NULL, E_WARNING, "%s", IBG(errmsg));
-	}
-}
-
-/* print php firebird module error and save it for fbird_errmsg() */
-void _php_fbird_module_error(const char *msg, ...)
-{
-	va_list ap;
-
-	va_start(ap, msg);
-
-	/* vsnprintf NUL terminates the buf and writes at most n-1 chars+NUL */
-	vsnprintf(IBG(errmsg), MAX_ERRMSG, msg, ap);
-	va_end(ap);
-
-	IBG(sql_code) = -999; /* no SQL error */
-
-	if (INI_BOOL("fbird.enable_exceptions")) {
-		zend_throw_exception(firebird_exception_ce, IBG(errmsg), IBG(sql_code));
-	} else {
-		php_error_docref(NULL, E_WARNING, "%s", IBG(errmsg));
-	}
-}
-
-typedef struct {
-	isc_db_handle *db_ptr;
-	zend_long tpb_len;
-	char *tpb_ptr;
-} ISC_TEB;
-
-
 /* Fill ib_link and trans with the correct database link and transaction. */
-void _php_fbird_get_link_trans(INTERNAL_FUNCTION_PARAMETERS,
-	zval *link_id, fbird_db_link **ib_link, fbird_transaction **trans)
-{
-	FBDEBUG("Transaction or database link?");
-	if (Z_RES_P(link_id)->type == le_trans) {
-		/* Transaction resource: make sure it refers to one link only, then
-		   fetch it; database link is stored in ib_trans->db_link[]. */
-		FBDEBUG("Type is le_trans");
-		*trans = (fbird_transaction *)zend_fetch_resource_ex(link_id, LE_TRANS, le_trans);
-		if ((*trans)->link_cnt > 1) {
-			_php_fbird_module_error("Link id is ambiguous: transaction spans multiple connections."
-				);
-			return;
-		}
-		*ib_link = (*trans)->db_link[0];
-		return;
-	}
-	FBDEBUG("Type is le_[p]link or id not found");
-	/* Database link resource, use default transaction. */
-	*trans = NULL;
-	*ib_link = (fbird_db_link *)zend_fetch_resource2_ex(link_id, LE_LINK, le_link, le_plink);
-}
-
-/* destructors ---------------------- */
-
-static void _php_fbird_commit_link(fbird_db_link *link)
-{
-	unsigned short i = 0, j;
-	fbird_tr_list *l;
-	fbird_event *e;
-	FBDEBUG("Checking transactions to close...");
-
-	for (l = link->tr_list; l != NULL; ++i) {
-		fbird_tr_list *p = l;
-		if (p->trans != 0) {
-			if (i == 0) {
-				/* Default transaction: commit via OO API */
-				if (p->trans->fbt_transaction != NULL) {
-					FBDEBUG("Committing default transaction via OO API...");
-					int res = fbt_commit(p->trans->fbt_transaction, IB_STATUS);
-					fbt_free(p->trans->fbt_transaction);
-					p->trans->fbt_transaction = NULL;
-					if (res) {
-						_php_fbird_error();
-					}
-				}
-				efree(p->trans); /* default transaction is not a registered resource: clean up */
-			} else {
-				/* Non-default transaction: rollback via OO API */
-				if (p->trans->fbt_transaction != NULL) {
-					FBDEBUG("Rolling back other transaction via OO API...");
-					int res = fbt_rollback(p->trans->fbt_transaction, IB_STATUS);
-					fbt_free(p->trans->fbt_transaction);
-					p->trans->fbt_transaction = NULL;
-					if (res) {
-						_php_fbird_error();
-					}
-				}
-				/* set this link pointer to NULL in the transaction */
-				for (j = 0; j < p->trans->link_cnt; ++j) {
-					if (p->trans->db_link[j] == link) {
-						p->trans->db_link[j] = NULL;
-						break;
-					}
-				}
-			}
-		}
-		l = l->next;
-		efree(p);
-	}
-	link->tr_list = NULL;
-
-	/*
-	 * IMPORTANT: _php_fbird_free_event() unlinks and efree()s the current node.
-	 * Never follow e->event_next after freeing e.
-	 */
-	for (e = link->event_head; e; ) {
-		fbird_event *next = e->event_next;
-		_php_fbird_free_event(e);
-		/* e is freed inside _php_fbird_free_event(); only use cached next */
-		e = next;
-	}
-}
-
-
-static void php_fbird_commit_link_rsrc(zend_resource *rsrc)
-{
-	fbird_db_link *link = (fbird_db_link *) rsrc->ptr;
-
-	_php_fbird_commit_link(link);
-}
-
-static void _php_fbird_close_link(zend_resource *rsrc)
-{
-	fbird_db_link *link = (fbird_db_link *) rsrc->ptr;
-
-	/* NULL pointer guard (Issue #55): In forked PHPStan workers, rsrc->ptr may be NULL
-	 * when inherited resource descriptors are destroyed during child process shutdown.
-	 * Accessing link->created_pid with NULL pointer causes SIGSEGV at si_addr=0x4. */
-	if (link == NULL) {
-		return;
-	}
-
-#ifndef PHP_WIN32
-	/* Fork-safety check (Issue #22, #36): Skip cleanup if we're in a forked child.
-	 * After pcntl_fork(), child inherits global state including master_instance
-	 * and connection handles. Attempting to close handles in child that were
-	 * created in parent causes segfault. Only the original process should
-	 * perform cleanup operations.
-	 *
-	 * Two-level check:
-	 * 1. Global init_pid - module-level fork detection
-	 * 2. Per-connection created_pid - connection-level fork detection */
-	pid_t current_pid = getpid();
-	if (IBG(init_pid) != 0 && current_pid != IBG(init_pid)) {
-		FBDEBUG("Skipping link cleanup in forked child process (global)");
-		IBG(num_links)--;
-		efree(link);
-		return;
-	}
-	if (link->created_pid != 0 && current_pid != link->created_pid) {
-		FBDEBUG("Skipping link cleanup in forked child process (per-connection)");
-		IBG(num_links)--;
-		efree(link);
-		return;
-	}
-#endif
-
-	/* Remove cache entry from EG(regular_list) to prevent UAF (Issue #35).
-	 *
-	 * CRITICAL: Skip EG() access during MSHUTDOWN (Issue #50, #51, #55).
-	 * During module shutdown, EG(regular_list) may already be destroyed. */
-	if (!IBG(in_mshutdown) &&
-		(link->hash_key[0] != '\0' || memcmp(link->hash_key, "\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0", 16) != 0)) {
-		zend_hash_str_del(&EG(regular_list), link->hash_key, sizeof(link->hash_key) - 1);
-		FBDEBUG("Removed cache entry for normal link");
-	}
-
-	_php_fbird_commit_link(link);
-
-	/* OO API Only: All connections use fbc_disconnect() */
-	if (link->fbc_connection != NULL) {
-		FBDEBUG("Closing normal link via OO API...");
-		fbc_disconnect(link->fbc_connection, IB_STATUS);
-		link->fbc_connection = NULL;
-		link->handle.ptr = 0;
-	}
-	IBG(num_links)--;
-	efree(link);
-}
-
-static void _php_fbird_close_plink(zend_resource *rsrc)
-{
-	fbird_db_link *link = (fbird_db_link *) rsrc->ptr;
-
-	/* NULL pointer guard (Issue #55): In forked PHPStan workers, rsrc->ptr may be NULL
-	 * when inherited resource descriptors are destroyed during child process shutdown.
-	 * Accessing link->created_pid with NULL pointer causes SIGSEGV at si_addr=0x4. */
-	if (link == NULL) {
-		return;
-	}
-
-#ifndef PHP_WIN32
-	/* Fork-safety check (Issue #22, #36): Skip cleanup if we're in a forked child.
-	 * Two-level check for both module-level and connection-level fork detection. */
-	pid_t current_pid = getpid();
-	if (IBG(init_pid) != 0 && current_pid != IBG(init_pid)) {
-		FBDEBUG("Skipping persistent link cleanup in forked child process (global)");
-		IBG(num_persistent)--;
-		IBG(num_links)--;
-		free(link);
-		return;
-	}
-	if (link->created_pid != 0 && current_pid != link->created_pid) {
-		FBDEBUG("Skipping persistent link cleanup in forked child process (per-connection)");
-		IBG(num_persistent)--;
-		IBG(num_links)--;
-		free(link);
-		return;
-	}
-#endif
-
-	/* Remove cache entries from both regular and persistent lists (Issue #35).
-	 * Persistent connections are cached in EG(persistent_list) with hash key.
-	 *
-	 * CRITICAL: Skip EG() access during MSHUTDOWN (Issue #50, #51).
-	 * During module shutdown, EG(regular_list) and EG(persistent_list) may already
-	 * be destroyed, causing SIGSEGV (exit code 139) if accessed. */
-	if (!IBG(in_mshutdown) &&
-		(link->hash_key[0] != '\0' || memcmp(link->hash_key, "\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0", 16) != 0)) {
-		zend_hash_str_del(&EG(regular_list), link->hash_key, sizeof(link->hash_key) - 1);
-		zend_hash_str_del(&EG(persistent_list), link->hash_key, sizeof(link->hash_key) - 1);
-		FBDEBUG("Removed cache entries for persistent link");
-	}
-
-	_php_fbird_commit_link(link);
-
-	/* OO API Only: All connections use fbc_disconnect() */
-	if (link->fbc_connection != NULL) {
-		FBDEBUG("Closing permanent link via OO API...");
-		fbc_disconnect(link->fbc_connection, IB_STATUS);
-		link->fbc_connection = NULL;
-		link->handle.ptr = 0;
-	}
-	IBG(num_persistent)--;
-	IBG(num_links)--;
-	free(link);
-}
-
-static void _php_fbird_free_trans(zend_resource *rsrc)
-{
-	fbird_transaction *trans = (fbird_transaction *)rsrc->ptr;
-	unsigned short i;
-
-	FBDEBUG("Cleaning up transaction resource...");
-
-#ifndef PHP_WIN32
-	/* Fork-safety check (Issue #22): Skip cleanup if we're in a forked child */
-	if (IBG(init_pid) != 0 && getpid() != IBG(init_pid)) {
-		FBDEBUG("Skipping transaction cleanup in forked child process");
-		efree(trans);
-		return;
-	}
-#endif
-
-	/* OO API Only: All transactions use fbt_rollback() */
-	if (trans->fbt_transaction != NULL) {
-		FBDEBUG("Rolling back unhandled OO API transaction...");
-		int res = fbt_rollback(trans->fbt_transaction, IB_STATUS);
-		fbt_free(trans->fbt_transaction);
-		trans->fbt_transaction = NULL;
-		trans->handle.ptr = 0;
-		if (res) {
-			_php_fbird_error();
-		}
-	}
-
-	/* now remove this transaction from all the connection-transaction lists */
-	for (i = 0; i < trans->link_cnt; ++i) {
-		if (trans->db_link[i] != NULL) {
-			fbird_tr_list **l;
-			for (l = &trans->db_link[i]->tr_list; *l != NULL; l = &(*l)->next) {
-				if ( (*l)->trans == trans) {
-					fbird_tr_list *p = *l;
-					*l = p->next;
-					efree(p);
-					break;
-				}
-			}
-		}
-	}
-	efree(trans);
-}
-
-#if FB_API_VER >= 40
-static void _php_fbird_free_batch(zend_resource *rsrc)
-{
-	fbird_batch *batch = (fbird_batch *)rsrc->ptr;
-
-	FBDEBUG("Cleaning up batch resource...");
-
-#ifndef PHP_WIN32
-	/* Fork-safety check (Issue #22): Skip cleanup if we're in a forked child */
-	if (IBG(init_pid) != 0 && getpid() != IBG(init_pid)) {
-		FBDEBUG("Skipping batch cleanup in forked child process");
-		/* Only free memory allocated by child, not Firebird handles */
-		if (batch->in_msg_buffer != NULL) {
-			efree(batch->in_msg_buffer);
-		}
-		efree(batch);
-		return;
-	}
-#endif
-
-	/* Cancel and close the batch if still open */
-	if (batch->fbbatch_wrapper != NULL) {
-		FBDEBUG("Canceling unexecuted batch...");
-		fbbatch_cancel(IBG(master_instance), batch->fbbatch_wrapper, IB_STATUS);
-		fbbatch_close(IBG(master_instance), batch->fbbatch_wrapper, IB_STATUS);
-		batch->fbbatch_wrapper = NULL;
-	}
-
-	/* Free the input message buffer if allocated */
-	if (batch->in_msg_buffer != NULL) {
-		efree(batch->in_msg_buffer);
-		batch->in_msg_buffer = NULL;
-	}
-
-	/* Release metadata reference if held */
-	if (batch->in_metadata != NULL) {
-		fbm_release(batch->in_metadata);
-		batch->in_metadata = NULL;
-	}
-
-	efree(batch);
-}
-#endif /* FB_API_VER >= 40 */
-
-/*
- * Custom INI display callback for password fields.
- * Displays "********" instead of the actual password value in phpinfo().
- * Note: This pattern is common across extensions; consider proposing for Zend API.
- */
 static PHP_INI_DISP(php_fbird_password_displayer_cb)
 {
 
@@ -1448,1349 +940,6 @@ int _php_fbird_attach_db(char **args, size_t *len, zend_long *largs, void **db)
     return SUCCESS;
 }
 
-static void _php_fbird_connect(INTERNAL_FUNCTION_PARAMETERS, int persistent)
-{
-	char *c, hash[16], *args[] = { NULL, NULL, NULL, NULL, NULL };
-	int i;
-	size_t len[] = { 0, 0, 0, 0, 0 };
-	zend_long largs[] = { 0, 0, 0 };
-	zend_long flags = 0;
-	PHP_MD5_CTX hash_context;
-	zend_resource new_index_ptr, *le;
-	void *db_handle = 0;
-	fbird_db_link *ib_link;
-
-	RESET_ERRMSG;
-
-	if (FAILURE == zend_parse_parameters(ZEND_NUM_ARGS(), "|ssssllsll",
-			&args[DB], &len[DB], &args[USER], &len[USER], &args[PASS], &len[PASS],
-			&args[CSET], &len[CSET], &largs[BUF], &largs[DLECT], &args[ROLE], &len[ROLE],
-			&largs[SYNC], &flags)) {
-		RETURN_FALSE;
-	}
-
-	/* restrict to the server/db in the .ini if in safe mode */
-	if (!len[DB] && (c = INI_STR("fbird.default_db"))) {
-		args[DB] = c;
-		len[DB] = strlen(c);
-	}
-	if (!len[USER] && (c = INI_STR("fbird.default_user"))) {
-		args[USER] = c;
-		len[USER] = strlen(c);
-	}
-	if (!len[PASS] && (c = INI_STR("fbird.default_password"))) {
-		args[PASS] = c;
-		len[PASS] = strlen(c);
-	}
-	if (!len[CSET] && (c = INI_STR("fbird.default_charset"))) {
-		args[CSET] = c;
-		len[CSET] = strlen(c);
-	}
-
-	/* don't want usernames and passwords floating around */
-	PHP_MD5Init(&hash_context);
-	for (i = 0; i < sizeof(args)/sizeof(char*); ++i) {
-		PHP_MD5Update(&hash_context,args[i],len[i]);
-	}
-	for (i = 0; i < sizeof(largs)/sizeof(zend_long); ++i) {
-		PHP_MD5Update(&hash_context,(char*)&largs[i],sizeof(zend_long));
-	}
-	PHP_MD5Final((unsigned char*)hash, &hash_context);
-
-	/* try to reuse a connection (skip if FBIRD_CONNECT_FORCE_NEW is set) */
-	if (!(flags & PHP_FBIRD_CONNECT_FORCE_NEW) &&
-			(le = zend_hash_str_find_ptr(&EG(regular_list), hash, sizeof(hash)-1)) != NULL) {
-		zend_resource *xlink;
-
-		if (le->type != le_index_ptr) {
-			RETURN_FALSE;
-		}
-
-		xlink = (zend_resource*) le->ptr;
-		if ((!persistent && xlink->type == le_link) || xlink->type == le_plink) {
-			if (IBG(default_link) != xlink) {
-				GC_ADDREF(xlink);
-				if (IBG(default_link)) {
-					zend_list_delete(IBG(default_link));
-				}
-				IBG(default_link) = xlink;
-			}
-			GC_ADDREF(xlink);
-			RETURN_RES(xlink);
-		} else {
-			zend_hash_str_del(&EG(regular_list), hash, sizeof(hash)-1);
-		}
-	}
-
-	/* ... or a persistent one */
-	do {
-		zend_long l;
-		static char info[] = { isc_info_base_level, isc_info_end };
-		char result[8];
-		ISC_STATUS status[20];
-
-		if ((le = zend_hash_str_find_ptr(&EG(persistent_list), hash, sizeof(hash)-1)) != NULL) {
-			if (le->type != le_plink) {
-				RETURN_FALSE;
-			}
-			/* check if connection has timed out */
-			ib_link = (fbird_db_link *) le->ptr;
-			if (!isc_database_info(status, &ib_link->handle.db, sizeof(info), info, sizeof(result), result)) {
-				RETVAL_RES(zend_register_resource(ib_link, le_plink));
-				break;
-			}
-			zend_hash_str_del(&EG(persistent_list), hash, sizeof(hash)-1);
-		}
-
-		/* no link found, so we have to open one */
-
-		if ((l = INI_INT("fbird.max_links")) != -1 && IBG(num_links) >= l) {
-			_php_fbird_module_error("Too many open links (%ld)", IBG(num_links));
-			RETURN_FALSE;
-		}
-
-		/* create the ib_link */
-		if (FAILURE == _php_fbird_attach_db(args, len, largs, &db_handle)) {
-			RETURN_FALSE;
-		}
-
-		/* use non-persistent if allowed number of persistent links is exceeded */
-		if (!persistent || ((l = INI_INT("fbird.max_persistent") != -1) && IBG(num_persistent) >= l)) {
-			ib_link = (fbird_db_link *) emalloc(sizeof(fbird_db_link));
-			RETVAL_RES(zend_register_resource(ib_link, le_link));
-		} else {
-			ib_link = (fbird_db_link *) malloc(sizeof(fbird_db_link));
-			if (!ib_link) {
-				RETURN_FALSE;
-			}
-
-			/* hash it up */
-			if (zend_register_persistent_resource(hash, sizeof(hash)-1, ib_link, le_plink) == NULL) {
-				free(ib_link);
-				RETURN_FALSE;
-			}
-			RETVAL_RES(zend_register_resource(ib_link, le_plink));
-			++IBG(num_persistent);
-		}
-		ib_link->handle.ptr = db_handle;
-		ib_link->dialect = largs[DLECT] ? (unsigned short)largs[DLECT] : SQL_DIALECT_CURRENT;
-		ib_link->tr_list = NULL;
-		ib_link->event_head = NULL;
-
-		ib_link->fbc_connection = (void *)(uintptr_t)IBG(status[ISC_STATUS_LENGTH - 1]);
-		IBG(status[ISC_STATUS_LENGTH - 1]) = 0;  /* Clear the temporary storage */
-
-		/* Store hash key for cache invalidation on close (Issue #35) */
-		memcpy(ib_link->hash_key, hash, sizeof(hash));
-
-		/* Store creation PID for fork-safety detection (Issue #36) */
-#ifndef PHP_WIN32
-		ib_link->created_pid = getpid();
-#else
-		ib_link->created_pid = 0;
-#endif
-
-		++IBG(num_links);
-	} while (0);
-
-	/* add it to the hash */
-	new_index_ptr.ptr = (void *) Z_RES_P(return_value);
-	new_index_ptr.type = le_index_ptr;
-	zend_hash_str_update_mem(&EG(regular_list), hash, sizeof(hash)-1,
-			(void *) &new_index_ptr, sizeof(zend_resource));
-	if (IBG(default_link)) {
-		zend_list_delete(IBG(default_link));
-	}
-	IBG(default_link) = Z_RES_P(return_value);
-	Z_TRY_ADDREF_P(return_value);
-	Z_TRY_ADDREF_P(return_value);
-}
-
-PHP_FUNCTION(fbird_connect)
-{
-	_php_fbird_connect(INTERNAL_FUNCTION_PARAM_PASSTHRU, 0);
-}
-
-PHP_FUNCTION(fbird_pconnect)
-{
-	_php_fbird_connect(INTERNAL_FUNCTION_PARAM_PASSTHRU, INI_INT("fbird.allow_persistent"));
-}
-
-/* Helper function for consolidated resource validation with proper error differentiation */
-static int _php_fbird_validate_link_resource(zend_resource *link_res, bool is_default_link, bool clear_default)
-{
-	if (link_res == NULL) {
-		return FAILURE;
-	}
-
-	/* Check resource type and validity directly */
-	if (link_res->type != le_link && link_res->type != le_plink) {
-		/* Wrong type - will be caught by zend_parse_parameters */
-		return FAILURE;
-	}
-
-	/* Check if resource pointer is valid */
-	if (link_res->ptr == NULL) {
-		/* Correct resource type but invalid/closed - generate warning */
-		php_error_docref(NULL, E_WARNING, "Supplied resource is not a valid database link resource");
-		if (clear_default && is_default_link) {
-			/* Thread-safe: Only clear if we were the default */
-			if (IBG(default_link) == link_res) {
-				IBG(default_link) = NULL;
-			}
-		}
-		return FAILURE;
-	}
-
-	return SUCCESS;
-}
-
-/* Helper function for thread-safe default link adoption */
-static void _php_fbird_adopt_new_default_link(zend_resource *closing_link)
-{
-	/* Only search if we're actually clearing the current default */
-	if (IBG(default_link) != closing_link) {
-		return;
-	}
-
-	/* For now, simply clear the default. Full adoption logic can be added in future enhancement.
-	 * This maintains existing behavior while providing the infrastructure for adoption. */
-	IBG(default_link) = NULL;
-}
-
-/* Helper function for optimized resource cleanup */
-static void _php_fbird_close_resource(zend_resource *link_res)
-{
-	/* For persistent connections, check reference count more carefully */
-	if (link_res->type == le_plink && GC_REFCOUNT(link_res) > 1) {
-		/* Multiple references exist - just decrease our refcount */
-		zend_list_delete(link_res);
-	} else {
-		/* Safe to close: either non-persistent or no other references */
-		zend_list_close(link_res);
-	}
-}
-
-PHP_FUNCTION(fbird_close)
-{
-	zval *link_arg = NULL;
-	zend_resource *link_res;
-	bool is_default_link = false;
-
-	RESET_ERRMSG;
-
-	if (zend_parse_parameters(ZEND_NUM_ARGS(), "|r!", &link_arg) == FAILURE) {
-		return;
-	}
-
-	/* Determine which link to close */
-	if (ZEND_NUM_ARGS() == 0 || link_arg == NULL) {
-		/* Default link path */
-		link_res = IBG(default_link);
-		is_default_link = true;
-
-		if (link_res == NULL) {
-			RETURN_FALSE;
-		}
-	} else {
-		/* Explicit link path */
-		link_res = Z_RES_P(link_arg);
-		is_default_link = (IBG(default_link) == link_res);
-	}
-
-	/* Single validation point - handles all validation efficiently */
-	if (_php_fbird_validate_link_resource(link_res, is_default_link, true) == FAILURE) {
-		RETURN_FALSE;
-	}
-
-	/* Handle default link management BEFORE closing resource
-	 * Special handling for explicit links that are also the default */
-	if (is_default_link) {
-		if (link_arg != NULL) {
-			/* When closing explicit link that's also default, only clear if
-			 * resource's reference count will drop to zero */
-			if (GC_REFCOUNT(link_res) <= 2) {
-				_php_fbird_adopt_new_default_link(link_res);
-			}
-		} else {
-			/* Default link path - always clear default */
-			_php_fbird_adopt_new_default_link(link_res);
-		}
-	}
-
-	/* Optimized resource cleanup */
-	_php_fbird_close_resource(link_res);
-
-	RETURN_TRUE;
-}
-
-PHP_FUNCTION(fbird_drop_db)
-{
-	zval *link_arg = NULL;
-	fbird_db_link *ib_link;
-	fbird_tr_list *l;
-	zend_resource *link_res;
-	int drop_result;
-
-	RESET_ERRMSG;
-
-	if (zend_parse_parameters(ZEND_NUM_ARGS(), "|r", &link_arg) == FAILURE) {
-		return;
-	}
-
-	if (ZEND_NUM_ARGS() == 0) {
-		link_res = IBG(default_link);
-		CHECK_LINK(link_res);
-		IBG(default_link) = NULL;
-	} else {
-		link_res = Z_RES_P(link_arg);
-	}
-
-	ib_link = (fbird_db_link *)zend_fetch_resource2(link_res, LE_LINK, le_link, le_plink);
-
-	if (!ib_link) {
-		RETURN_FALSE;
-	}
-
-	/* OO API Only: All connections use fbc_drop_database() */
-	if (ib_link->fbc_connection != NULL) {
-		FBDEBUG("Dropping database via OO API...");
-		drop_result = fbc_drop_database(ib_link->fbc_connection, IB_STATUS);
-		if (drop_result != 0) {
-			_php_fbird_error();
-			RETURN_FALSE;
-		}
-		/* fbc_drop_database() already frees the connection wrapper */
-		ib_link->fbc_connection = NULL;
-		ib_link->handle.ptr = 0;
-	}
-
-	/* drop_database() doesn't invalidate the transaction handles */
-	for (l = ib_link->tr_list; l != NULL; l = l->next) {
-		if (l->trans != NULL) {
-			l->trans->handle.ptr = 0;
-			l->trans->fbt_transaction = NULL;
-		}
-	}
-
-	zend_list_delete(link_res);
-
-	RETURN_TRUE;
-}
-
-
-#define TPB_MAX_SIZE 2048
-
-void _php_fbird_populate_trans(zend_long trans_argl, zend_long trans_timeout, char *last_tpb, unsigned short *len)
-{
-	/* No explicit flags: leave TPB empty so Firebird uses its defaults. */
-	if (trans_argl == PHP_FBIRD_DEFAULT) {
-		*len = 0;
-		return;
-	}
-
-	/*
-	 * Use IXpbBuilder-based TPB construction via OO API.
-	 * This replaces manual byte array construction with the modern
-	 * Firebird 3.0+ builder pattern for cleaner, type-safe TPB generation.
-	 *
-	 * See: fbxpb_build_tpb() in firebird_utils.cpp
-	 */
-	unsigned int buffer_length = 0;
-	ISC_STATUS local_status[ISC_STATUS_LENGTH] = {0};
-
-	unsigned char *tpb_buffer = fbxpb_build_tpb(
-		IBG(master_instance),
-		trans_argl,
-		trans_timeout,
-		&buffer_length,
-		local_status
-	);
-
-	if (tpb_buffer == NULL) {
-		/* Fallback: if OO API fails, report error and return empty TPB */
-		php_error_docref(NULL, E_WARNING, "DEBUG: fbxpb_build_tpb returned NULL, local_status[1]=%ld", (long)local_status[1]);
-		*len = 0;
-		return;
-	}
-
-	/* Copy to caller's buffer (limited by TPB_MAX_SIZE) */
-	if (buffer_length > TPB_MAX_SIZE) {
-		buffer_length = TPB_MAX_SIZE;
-	}
-	memcpy(last_tpb, tpb_buffer, buffer_length);
-	*len = (unsigned short) buffer_length;
-
-	/* Free the OO API allocated buffer */
-	fbxpb_free_tpb(tpb_buffer);
-}
-
-void _php_fbird_populate_trans_from_array(zval *options, zend_long *trans_timeout, char *last_tpb, unsigned short *len)
-{
-	unsigned char *p = (unsigned char *) last_tpb;
-	unsigned char *end = p + TPB_MAX_SIZE;
-	zval *tmp;
-
-	/* TPB version */
-	*p++ = isc_tpb_version3;
-
-	/* access mode */
-	if ((tmp = zend_hash_str_find(Z_ARRVAL_P(options), "access_mode", sizeof("access_mode") - 1)) != NULL) {
-		if (Z_TYPE_P(tmp) == IS_LONG) {
-			if (Z_LVAL_P(tmp) & PHP_FBIRD_READ) {
-				*p++ = isc_tpb_read;
-			} else if (Z_LVAL_P(tmp) & PHP_FBIRD_WRITE) {
-				*p++ = isc_tpb_write;
-			}
-		}
-	}
-
-	/* isolation level */
-	if ((tmp = zend_hash_str_find(Z_ARRVAL_P(options), "isolation", sizeof("isolation") - 1)) != NULL) {
-		if (Z_TYPE_P(tmp) == IS_LONG) {
-			zend_long iso = Z_LVAL_P(tmp);
-			if (iso & PHP_FBIRD_COMMITTED) {
-				*p++ = isc_tpb_read_committed;
-				if (iso & PHP_FBIRD_REC_VERSION) {
-					*p++ = isc_tpb_rec_version;
-				} else if (iso & PHP_FBIRD_REC_NO_VERSION) {
-					*p++ = isc_tpb_no_rec_version;
-				}
-			} else if (iso & PHP_FBIRD_CONSISTENCY) {
-				*p++ = isc_tpb_consistency;
-			} else if (iso & PHP_FBIRD_CONCURRENCY) {
-				*p++ = isc_tpb_concurrency;
-			}
-		}
-	}
-
-	/* lock resolution */
-	if ((tmp = zend_hash_str_find(Z_ARRVAL_P(options), "lock_resolution", sizeof("lock_resolution") - 1)) != NULL) {
-		if (Z_TYPE_P(tmp) == IS_LONG) {
-			zend_long res = Z_LVAL_P(tmp);
-			if (res & PHP_FBIRD_NOWAIT) {
-				*p++ = isc_tpb_nowait;
-			} else if (res & PHP_FBIRD_WAIT) {
-				*p++ = isc_tpb_wait;
-			}
-		}
-	} else if ((tmp = zend_hash_str_find(Z_ARRVAL_P(options), "wait", sizeof("wait") - 1)) != NULL) {
-		/* BC for simpler key 'wait' => true/false? No, prefer explicit constants. */
-		/* Let's stick to RFC constants. */
-	}
-
-	/* lock_timeout */
-	if ((tmp = zend_hash_str_find(Z_ARRVAL_P(options), "lock_timeout", sizeof("lock_timeout") - 1)) != NULL) {
-		if (Z_TYPE_P(tmp) == IS_LONG) {
-			zend_long timeout = Z_LVAL_P(tmp);
-			if (timeout > 0 && timeout <= 0x7FFF) {
-				/* If we have a timeout, we implicitly need WAIT */
-				/* Check if user already set NOWAIT? If conflicting, timeout usually ignored or error.
-				   Firebird: isc_tpb_wait is required for isc_tpb_lock_timeout.
-				   If user didn't set resolution, we add isc_tpb_wait.
-				   But we can't easily check if we already added it in the stream without parsing back or tracking state.
-				   Let's assume smart usage or add isc_tpb_wait if not added?
-				   Actually, duplications in TPB are usually fine or last one wins?
-				   Let's just append isc_tpb_lock_timeout. */
-				/* Actually, standard implementation in populate_trans handles it by nesting. */
-
-				*p++ = isc_tpb_lock_timeout;
-				*p++ = (unsigned char) sizeof(ISC_SHORT);
-				/* VAX/Firebird little-endian order */
-				*p++ = (unsigned char) (timeout & 0xff);
-				*p++ = (unsigned char) ((timeout >> 8) & 0xff);
-				*trans_timeout = timeout;
-			}
-		}
-	}
-
-	/* read_consistency (Firebird 4.0+) */
-	if ((tmp = zend_hash_str_find(Z_ARRVAL_P(options), "read_consistency", sizeof("read_consistency") - 1)) != NULL) {
-		if (zend_is_true(tmp)) {
-			*p++ = isc_tpb_read_consistency;
-			*p++ = 1;
-		}
-	}
-
-	/* tables reservation (Firebird 1.5+) */
-	if ((tmp = zend_hash_str_find(Z_ARRVAL_P(options), "tables", sizeof("tables") - 1)) != NULL) {
-		if (Z_TYPE_P(tmp) == IS_ARRAY) {
-			zval *table_val;
-			zend_string *table_name;
-
-			ZEND_HASH_FOREACH_STR_KEY_VAL(Z_ARRVAL_P(tmp), table_name, table_val) {
-				if (table_name && Z_TYPE_P(table_val) == IS_LONG) {
-					zend_long lock_mode = Z_LVAL_P(table_val);
-					unsigned char tlen = (unsigned char) ZSTR_LEN(table_name);
-
-					// Basic Direction (Read/Write)
-					if (lock_mode & PHP_FBIRD_LOCK_WRITE) {
-						if (p < end) *p++ = isc_tpb_lock_write;
-					} else if (lock_mode & PHP_FBIRD_LOCK_READ) {
-						if (p < end) *p++ = isc_tpb_lock_read;
-					} else {
-						// Default to WRITE if only mode bits are set (common assumption) or SKIP?
-						// Let's skip if no direction is set to avoid invalid TPB
-						continue;
-					}
-
-					// Table Name
-					if (p + 1 + tlen < end) {
-						*p++ = tlen;
-						memcpy(p, ZSTR_VAL(table_name), tlen);
-						p += tlen;
-					} else {
-						// Buffer overflow prevention
-						php_error_docref(NULL, E_WARNING, "TPB buffer too small for table '%s'", ZSTR_VAL(table_name));
-						break;
-					}
-
-					// Lock Mode (Shared/Protected/Exclusive)
-					if (lock_mode & PHP_FBIRD_LOCK_PROTECTED) {
-						if (p < end) *p++ = isc_tpb_protected;
-					} else if (lock_mode & PHP_FBIRD_LOCK_EXCLUSIVE) {
-						if (p < end) *p++ = isc_tpb_exclusive;
-					} else if (lock_mode & PHP_FBIRD_LOCK_SHARED) {
-						if (p < end) *p++ = isc_tpb_shared;
-					} else {
-						// Default to SHARED if no mode specified
-						if (p < end) *p++ = isc_tpb_shared;
-					}
-				}
-			} ZEND_HASH_FOREACH_END();
-		}
-	}
-
-	*len = (unsigned short) (p - (unsigned char *) last_tpb);
-}
-
-PHP_FUNCTION(fbird_trans_start)
-{
-	zval *link_arg = NULL, *options_arg = NULL;
-	fbird_db_link *ib_link;
-	fbird_transaction *ib_trans;
-	void *tr_handle = 0;
-	ISC_STATUS result;
-	char last_tpb[TPB_MAX_SIZE];
-	unsigned short tpb_len = 0;
-	zend_long trans_timeout = 0;
-
-	RESET_ERRMSG;
-
-	if (zend_parse_parameters(ZEND_NUM_ARGS(), "|ra", &link_arg, &options_arg) == FAILURE) {
-		return;
-	}
-
-	/* If first arg is array, it's options, and use default link */
-	if (link_arg && Z_TYPE_P(link_arg) == IS_ARRAY) {
-		options_arg = link_arg;
-		link_arg = NULL;
-	}
-
-	if (link_arg) {
-		ib_link = (fbird_db_link *)zend_fetch_resource2_ex(link_arg, LE_LINK, le_link, le_plink);
-	} else {
-		ib_link = (fbird_db_link *)zend_fetch_resource2(IBG(default_link), LE_LINK, le_link, le_plink);
-	}
-
-	if (!ib_link) {
-		RETURN_FALSE;
-	}
-
-	if (options_arg) {
-		_php_fbird_populate_trans_from_array(options_arg, &trans_timeout, last_tpb, &tpb_len);
-	} else {
-		/* Default transaction parameters */
-		zend_long trans_argl = IBG(default_trans_params);
-		trans_timeout = IBG(default_lock_timeout);
-		_php_fbird_populate_trans(trans_argl, trans_timeout, last_tpb, &tpb_len);
-	}
-
-	/* OO API Only: All connections use fbt_start() */
-	if (ib_link->fbc_connection == NULL) {
-		_php_fbird_module_error("Connection has no OO API handle");
-		RETURN_FALSE;
-	}
-
-	void* attachment = fbc_get_attachment(ib_link->fbc_connection);
-	if (attachment == NULL) {
-		_php_fbird_module_error("Failed to get attachment from OO API connection");
-		RETURN_FALSE;
-	}
-
-	ib_trans = (fbird_transaction *) safe_emalloc(1-1, sizeof(fbird_db_link *), sizeof(fbird_transaction));
-	ib_trans->fbt_transaction = fbt_start(
-		IBG(master_instance),
-		attachment,
-		tpb_len,
-		tpb_len > 0 ? (const unsigned char*)last_tpb : NULL,
-		IB_STATUS
-	);
-
-	if (ib_trans->fbt_transaction == NULL) {
-		efree(ib_trans);
-		_php_fbird_error();
-		RETURN_FALSE;
-	}
-
-	ib_trans->handle.ptr = fbt_get_handle(ib_trans->fbt_transaction);
-	ib_trans->link_cnt = 1;
-	ib_trans->affected_rows = 0;
-	ib_trans->db_link[0] = ib_link;
-
-	/* the first item in the connection-transaction list is reserved for the default transaction */
-	if (ib_link->tr_list == NULL) {
-		ib_link->tr_list = (fbird_tr_list *) emalloc(sizeof(fbird_tr_list));
-		ib_link->tr_list->trans = NULL;
-		ib_link->tr_list->next = NULL;
-	}
-
-	/* link the transaction into the connection-transaction list */
-	fbird_tr_list **l;
-	for (l = &ib_link->tr_list; *l != NULL; l = &(*l)->next);
-	*l = (fbird_tr_list *) emalloc(sizeof(fbird_tr_list));
-	(*l)->trans = ib_trans;
-	(*l)->next = NULL;
-
-	RETVAL_RES(zend_register_resource(ib_trans, le_trans));
-	Z_TRY_ADDREF_P(return_value);
-}
-
-static void _php_fbird_exec_savepoint(INTERNAL_FUNCTION_PARAMETERS, const char *format)
-{
-	zval *trans_arg = NULL;
-	char *name;
-	size_t name_len;
-	fbird_transaction *trans;
-	char *query;
-	int len;
-
-	RESET_ERRMSG;
-
-	if (FAILURE == zend_parse_parameters(ZEND_NUM_ARGS(), "rs", &trans_arg, &name, &name_len)) {
-		return;
-	}
-
-	if (name_len == 0 || name_len > 31) { // Max identifier length
-		php_error_docref(NULL, E_WARNING, "Invalid savepoint name (length must be 1-31 bytes)");
-		RETURN_FALSE;
-	}
-
-	trans = (fbird_transaction *)zend_fetch_resource_ex(trans_arg, LE_TRANS, le_trans);
-	if (!trans) {
-		RETURN_FALSE;
-	}
-
-	/* Check if transaction involves exactly one connection */
-	if (trans->link_cnt > 1) {
-		_php_fbird_module_error("Savepoints not supported for multi-database transactions");
-		RETURN_FALSE;
-	}
-
-	len = spprintf(&query, 0, format, name);
-
-	/* OO API Only: All transactions use fbs_prepare() + fbs_execute() */
-	if (trans->fbt_transaction == NULL) {
-		_php_fbird_module_error("Transaction has no OO API handle");
-		efree(query);
-		RETURN_FALSE;
-	}
-
-	fbird_db_link *link = trans->db_link[0];
-	void *attachment = NULL;
-	void *transaction_ptr = NULL;
-	void *stmt = NULL;
-
-	FBDEBUG("OO API: Executing savepoint statement via fbs_prepare/execute");
-
-	/* Get attachment from connection */
-	if (!link->fbc_connection) {
-		_php_fbird_module_error("OO API transaction without OO API connection");
-		efree(query);
-		RETURN_FALSE;
-	}
-	attachment = fbc_get_attachment(link->fbc_connection);
-	if (!attachment) {
-		_php_fbird_module_error("Failed to get attachment from connection");
-		efree(query);
-		RETURN_FALSE;
-	}
-
-	/* Get transaction handle */
-	transaction_ptr = fbt_get_handle(trans->fbt_transaction);
-	if (!transaction_ptr) {
-		_php_fbird_module_error("Failed to get transaction handle");
-		efree(query);
-		RETURN_FALSE;
-	}
-
-	/* Prepare the savepoint statement */
-	stmt = fbs_prepare(IBG(master_instance), attachment, transaction_ptr,
-		query, (unsigned)len, SQL_DIALECT_CURRENT, IB_STATUS);
-	if (!stmt) {
-		_php_fbird_error();
-		efree(query);
-		RETURN_FALSE;
-	}
-
-	/* Execute the savepoint statement (no input/output parameters) */
-	if (!fbs_execute(IBG(master_instance), stmt, transaction_ptr,
-			NULL, NULL, NULL, NULL, IB_STATUS)) {
-		_php_fbird_error();
-		fbs_free(stmt, IB_STATUS);
-		efree(query);
-		RETURN_FALSE;
-	}
-
-	/* Free the statement */
-	fbs_free(stmt, IB_STATUS);
-	efree(query);
-	RETURN_TRUE;
-}
-
-PHP_FUNCTION(fbird_savepoint)
-{
-	_php_fbird_exec_savepoint(INTERNAL_FUNCTION_PARAM_PASSTHRU, "SAVEPOINT %s");
-}
-
-PHP_FUNCTION(fbird_rollback_savepoint)
-{
-	_php_fbird_exec_savepoint(INTERNAL_FUNCTION_PARAM_PASSTHRU, "ROLLBACK TO SAVEPOINT %s");
-}
-
-PHP_FUNCTION(fbird_release_savepoint)
-{
-	_php_fbird_exec_savepoint(INTERNAL_FUNCTION_PARAM_PASSTHRU, "RELEASE SAVEPOINT %s");
-}
-
-PHP_FUNCTION(fbird_trans_info)
-{
-	zval *trans_arg;
-	fbird_transaction *trans;
-	char tpb[] = {
-		isc_info_tra_id,
-		isc_info_tra_isolation,
-		isc_info_tra_lock_timeout,
-		isc_info_tra_access
-	};
-	char res_buf[128];
-	char *p = res_buf;
-
-	RESET_ERRMSG;
-
-	if (FAILURE == zend_parse_parameters(ZEND_NUM_ARGS(), "r", &trans_arg)) {
-		RETURN_FALSE;
-	}
-
-	trans = (fbird_transaction *)zend_fetch_resource_ex(trans_arg, LE_TRANS, le_trans);
-	if (!trans) {
-		RETURN_FALSE;
-	}
-
-	/*
-	 * Firebird 3.0+ OO API
-	 *
-	 * OO API transactions do not have a valid legacy isc_tr_handle.
-	 * For OO API transactions, trans->handle.ptr stores the raw ITransaction*
-	 * (from fbt_get_handle()).
-	 */
-	if (trans->fbt_transaction != NULL) {
-		if (trans->handle.ptr == NULL) {
-			_php_fbird_module_error("Transaction has no valid OO API handle");
-			RETURN_FALSE;
-		}
-
-		if (fbt_get_info(
-				IBG(master_instance),
-				trans->handle.ptr,
-				sizeof(tpb),
-				(const unsigned char*)tpb,
-				sizeof(res_buf),
-				(unsigned char*)res_buf,
-				IB_STATUS
-			) == 0) {
-			_php_fbird_error();
-			RETURN_FALSE;
-		}
-	} else {
-		_php_fbird_module_error("Transaction has no OO API handle");
-		RETURN_FALSE;
-	}
-
-	array_init(return_value);
-
-	while (*p != isc_info_end && p < res_buf + sizeof(res_buf)) {
-		unsigned char item = *p++;
-		unsigned short len = (unsigned short)isc_vax_integer(p, 2);
-		p += 2;
-
-		switch (item) {
-			case isc_info_tra_id:
-				add_assoc_long(return_value, "id", isc_vax_integer(p, len));
-				break;
-			case isc_info_tra_isolation:
-				if (len > 0) {
-					unsigned char iso = *p;
-					switch(iso) {
-						case isc_info_tra_consistency: add_assoc_string(return_value, "isolation", "CONSISTENCY"); break;
-						case isc_info_tra_concurrency: add_assoc_string(return_value, "isolation", "CONCURRENCY"); break;
-						case isc_info_tra_read_committed: add_assoc_string(return_value, "isolation", "READ_COMMITTED"); break;
-						default: add_assoc_long(return_value, "isolation_raw", iso); break;
-					}
-				}
-				break;
-			case isc_info_tra_lock_timeout:
-				add_assoc_long(return_value, "lock_timeout", isc_vax_integer(p, len));
-				break;
-			case isc_info_tra_access:
-				if (len > 0) {
-					add_assoc_string(return_value, "access_mode", (*p == isc_info_tra_readonly) ? "READ_ONLY" : "READ_WRITE");
-				}
-				break;
-			default:
-				/* Ignore unknown info items */
-				break;
-		}
-		p += len;
-	}
-
-	/* Add internal state tracking if possible, or just what API returned */
-	/* Since we don't track STATE in struct, we infer it is ACTIVE if valid resource */
-	add_assoc_string(return_value, "state", "ACTIVE");
-}
-
-PHP_FUNCTION(fbird_connection_info)
-{
-	zval *link_arg = NULL;
-	fbird_db_link *ib_link;
-	char info_items[] = {
-		isc_info_reads,
-		isc_info_writes,
-		isc_info_fetches,
-		isc_info_marks,
-		isc_info_page_size,
-		isc_info_num_buffers,
-		isc_info_current_memory,
-		isc_info_max_memory,
-		isc_info_allocation,
-		isc_info_attachment_id,
-		isc_info_ods_version,
-		isc_info_ods_minor_version,
-		isc_info_db_sql_dialect,
-		isc_info_end
-	};
-	char res_buf[512];
-	char *p;
-	ISC_STATUS status[ISC_STATUS_LENGTH];
-
-	RESET_ERRMSG;
-
-	if (zend_parse_parameters(ZEND_NUM_ARGS(), "|r!", &link_arg) == FAILURE) {
-		return;
-	}
-
-	if (link_arg == NULL) {
-		ib_link = (fbird_db_link *)zend_fetch_resource2(IBG(default_link), LE_LINK, le_link, le_plink);
-	} else {
-		ib_link = (fbird_db_link *)zend_fetch_resource2_ex(link_arg, LE_LINK, le_link, le_plink);
-	}
-
-	if (!ib_link) {
-		RETURN_FALSE;
-	}
-
-	/* Use OO API for connections that have fbc_connection */
-	if (ib_link->fbc_connection != NULL) {
-		void* attachment = fbc_get_attachment(ib_link->fbc_connection);
-		if (attachment == NULL) {
-			_php_fbird_module_error("Failed to get attachment from OO API connection");
-			RETURN_FALSE;
-		}
-
-		if (!fbc_get_info(IBG(master_instance), attachment,
-				sizeof(info_items), (const unsigned char*)info_items,
-				sizeof(res_buf), (unsigned char*)res_buf, IB_STATUS)) {
-			_php_fbird_error();
-			RETURN_FALSE;
-		}
-	} else {
-		/* Fallback to legacy API for connections without OO API handle */
-		if (isc_database_info(status, &ib_link->handle.db, sizeof(info_items), info_items,
-				sizeof(res_buf), res_buf)) {
-			memcpy(IB_STATUS, status, sizeof(status));
-			_php_fbird_error();
-			RETURN_FALSE;
-		}
-	}
-
-	array_init(return_value);
-	p = res_buf;
-
-	while (*p != isc_info_end && p < res_buf + sizeof(res_buf)) {
-		unsigned char item = *p++;
-		unsigned short len = (unsigned short)isc_vax_integer(p, 2);
-		p += 2;
-
-		switch (item) {
-			case isc_info_reads:
-				add_assoc_long(return_value, "reads", isc_vax_integer(p, len));
-				break;
-			case isc_info_writes:
-				add_assoc_long(return_value, "writes", isc_vax_integer(p, len));
-				break;
-			case isc_info_fetches:
-				add_assoc_long(return_value, "fetches", isc_vax_integer(p, len));
-				break;
-			case isc_info_marks:
-				add_assoc_long(return_value, "marks", isc_vax_integer(p, len));
-				break;
-			case isc_info_page_size:
-				add_assoc_long(return_value, "page_size", isc_vax_integer(p, len));
-				break;
-			case isc_info_num_buffers:
-				add_assoc_long(return_value, "num_buffers", isc_vax_integer(p, len));
-				break;
-			case isc_info_current_memory:
-				add_assoc_long(return_value, "current_memory", isc_vax_integer(p, len));
-				break;
-			case isc_info_max_memory:
-				add_assoc_long(return_value, "max_memory", isc_vax_integer(p, len));
-				break;
-			case isc_info_allocation:
-				add_assoc_long(return_value, "allocation", isc_vax_integer(p, len));
-				break;
-			case isc_info_attachment_id:
-				add_assoc_long(return_value, "attachment_id", isc_vax_integer(p, len));
-				break;
-			case isc_info_ods_version:
-				add_assoc_long(return_value, "ods_version", isc_vax_integer(p, len));
-				break;
-			case isc_info_ods_minor_version:
-				add_assoc_long(return_value, "ods_minor_version", isc_vax_integer(p, len));
-				break;
-			case isc_info_db_sql_dialect:
-				add_assoc_long(return_value, "sql_dialect", isc_vax_integer(p, len));
-				break;
-			case isc_info_truncated:
-				add_assoc_bool(return_value, "truncated", 1);
-				return;
-			default:
-				break;
-		}
-		p += len;
-	}
-}
-
-PHP_FUNCTION(fbird_trans)
-{
-	int i, argn = ZEND_NUM_ARGS();
-	unsigned short link_cnt = 0, tpb_len = 0;
-	char last_tpb[TPB_MAX_SIZE];
-	fbird_db_link **ib_link = NULL;
-	fbird_transaction *ib_trans;
-	void *tr_handle = 0;
-	ISC_STATUS result = 0;
-
-	RESET_ERRMSG;
-
-	/* (1+argn) is an upper bound for the number of links this trans connects to */
-	ib_link = (fbird_db_link **) safe_emalloc(sizeof(fbird_db_link *),1+argn,0);
-
-	if (argn > 0) {
-		zend_long trans_argl = 0;
-		zend_long trans_timeout = 0;
-		char *tpb;
-		ISC_TEB *teb;
-		zval *args = NULL;
-
-		if (zend_parse_parameters(argn, "+", &args, &argn) == FAILURE) {
-			efree(ib_link);
-			RETURN_FALSE;
-		}
-
-		teb = (ISC_TEB *) safe_emalloc(sizeof(ISC_TEB),argn,0);
-		tpb = (char *) safe_emalloc(TPB_MAX_SIZE,argn,0);
-
-		/* enumerate all the arguments: assume every non-resource argument
-		   specifies modifiers for the link ids that follow it */
-		for (i = 0; i < argn; ++i) {
-
-			if (Z_TYPE(args[i]) == IS_RESOURCE) {
-
-				if ((ib_link[link_cnt] = (fbird_db_link *)zend_fetch_resource2_ex(&args[i], LE_LINK, le_link, le_plink)) == NULL) {
-					efree(teb);
-					efree(tpb);
-					efree(ib_link);
-					RETURN_FALSE;
-				}
-
-				/* copy the most recent modifier string into tbp[] */
-				memcpy(&tpb[TPB_MAX_SIZE * link_cnt], last_tpb, TPB_MAX_SIZE);
-
-				/* add a database handle to the TEB with the most recently specified set of modifiers */
-				teb[link_cnt].db_ptr = &ib_link[link_cnt]->handle.db;
-				teb[link_cnt].tpb_len = tpb_len;
-				teb[link_cnt].tpb_ptr = &tpb[TPB_MAX_SIZE * link_cnt];
-
-				++link_cnt;
-
-			} else {
-
-				tpb_len = 0;
-
-				convert_to_long_ex(&args[i]);
-				trans_argl = Z_LVAL(args[i]);
-				if (trans_argl != PHP_FBIRD_DEFAULT) {
-					// Skip conflicting parameters
-					if (PHP_FBIRD_NOWAIT != (trans_argl & PHP_FBIRD_NOWAIT) && PHP_FBIRD_WAIT == (trans_argl & PHP_FBIRD_WAIT)) {
-						if (PHP_FBIRD_LOCK_TIMEOUT == (trans_argl & PHP_FBIRD_LOCK_TIMEOUT)) {
-							if((i + 1 < argn) && (Z_TYPE(args[i + 1]) == IS_LONG)){
-								i++;
-								convert_to_long_ex(&args[i]);
-								trans_timeout = Z_LVAL(args[i]);
-							} else {
-								php_error_docref(NULL, E_WARNING, "FBIRD_LOCK_TIMEOUT expects next argument to be timeout value");
-							}
-						}
-					}
-					_php_fbird_populate_trans(trans_argl, trans_timeout, last_tpb, &tpb_len);
-				}
-			}
-		}
-
-		if (link_cnt > 0) {
-			/* OO API Only: Verify all connections have OO API handles */
-			for (int j = 0; j < link_cnt; j++) {
-				if (ib_link[j]->fbc_connection == NULL) {
-					efree(tpb);
-					efree(teb);
-					efree(ib_link);
-					_php_fbird_module_error("Connection %d has no OO API handle", j);
-					RETURN_FALSE;
-				}
-			}
-
-			if (link_cnt == 1) {
-				/* Single OO API connection - use fbt_start */
-				void* attachment = fbc_get_attachment(ib_link[0]->fbc_connection);
-				if (attachment == NULL) {
-					efree(tpb);
-					efree(teb);
-					efree(ib_link);
-					_php_fbird_module_error("Failed to get attachment from OO API connection");
-					RETURN_FALSE;
-				}
-
-				void* oo_trans = fbt_start(
-					IBG(master_instance),
-					attachment,
-					teb[0].tpb_len,
-					teb[0].tpb_len > 0 ? (const unsigned char*)teb[0].tpb_ptr : NULL,
-					IB_STATUS
-				);
-
-				if (oo_trans == NULL) {
-					efree(tpb);
-					efree(teb);
-					efree(ib_link);
-					_php_fbird_error();
-					RETURN_FALSE;
-				}
-
-				tr_handle = fbt_get_handle(oo_trans);
-
-				/* Allocate and register transaction with OO API wrapper */
-				ib_trans = (fbird_transaction *) safe_emalloc(link_cnt-1, sizeof(fbird_db_link *), sizeof(fbird_transaction));
-				ib_trans->handle.ptr = tr_handle;
-				ib_trans->link_cnt = link_cnt;
-				ib_trans->affected_rows = 0;
-				ib_trans->fbt_transaction = oo_trans;
-
-				efree(tpb);
-				efree(teb);
-				goto register_trans;
-			} else {
-				/* Multi-database OO API transactions not yet supported */
-				efree(tpb);
-				efree(teb);
-				efree(ib_link);
-				_php_fbird_module_error("Multi-database transactions with OO API connections not yet supported");
-				RETURN_FALSE;
-			}
-		}
-
-		efree(tpb);
-		efree(teb);
-	}
-
-	if (link_cnt == 0) {
-		link_cnt = 1;
-		if ((ib_link[0] = (fbird_db_link *)zend_fetch_resource2(IBG(default_link), LE_LINK, le_link, le_plink)) == NULL) {
-			efree(ib_link);
-			RETURN_FALSE;
-		}
-		/* OO API Only: All connections must have OO API handle */
-		if (ib_link[0]->fbc_connection == NULL) {
-			efree(ib_link);
-			_php_fbird_module_error("Connection has no OO API handle");
-			RETURN_FALSE;
-		}
-
-		void* attachment = fbc_get_attachment(ib_link[0]->fbc_connection);
-		if (attachment == NULL) {
-			efree(ib_link);
-			_php_fbird_module_error("Failed to get attachment from OO API connection");
-			RETURN_FALSE;
-		}
-
-		void* oo_trans = fbt_start(
-			IBG(master_instance),
-			attachment,
-			tpb_len,
-			tpb_len > 0 ? (const unsigned char*)last_tpb : NULL,
-			IB_STATUS
-		);
-		if (oo_trans == NULL) {
-			_php_fbird_error();
-			efree(ib_link);
-			RETURN_FALSE;
-		}
-		tr_handle = fbt_get_handle(oo_trans);
-
-		/* Allocate and register transaction with OO API wrapper */
-		ib_trans = (fbird_transaction *) safe_emalloc(link_cnt-1, sizeof(fbird_db_link *), sizeof(fbird_transaction));
-		ib_trans->handle.ptr = tr_handle;
-		ib_trans->link_cnt = link_cnt;
-		ib_trans->affected_rows = 0;
-		ib_trans->fbt_transaction = oo_trans;
-	}
-
-register_trans:
-	for (i = 0; i < link_cnt; ++i) {
-		fbird_tr_list **l;
-		ib_trans->db_link[i] = ib_link[i];
-
-		/* the first item in the connection-transaction list is reserved for the default transaction */
-		if (ib_link[i]->tr_list == NULL) {
-			ib_link[i]->tr_list = (fbird_tr_list *) emalloc(sizeof(fbird_tr_list));
-			ib_link[i]->tr_list->trans = NULL;
-			ib_link[i]->tr_list->next = NULL;
-		}
-
-		/* link the transaction into the connection-transaction list */
-		for (l = &ib_link[i]->tr_list; *l != NULL; l = &(*l)->next);
-		*l = (fbird_tr_list *) emalloc(sizeof(fbird_tr_list));
-		(*l)->trans = ib_trans;
-		(*l)->next = NULL;
-	}
-	efree(ib_link);
-	RETVAL_RES(zend_register_resource(ib_trans, le_trans));
-	Z_TRY_ADDREF_P(return_value);
-}
-
-int _php_fbird_def_trans(fbird_db_link *ib_link, fbird_transaction **trans)
-{
-	if (ib_link == NULL) {
-		php_error_docref(NULL, E_WARNING, "Invalid database link");
-		return FAILURE;
-	}
-
-	/* the first item in the connection-transaction list is reserved for the default transaction */
-	if (ib_link->tr_list == NULL) {
-		ib_link->tr_list = (fbird_tr_list *) emalloc(sizeof(fbird_tr_list));
-		ib_link->tr_list->trans = NULL;
-		ib_link->tr_list->next = NULL;
-	}
-
-	if (*trans == NULL) {
-		fbird_transaction *tr = ib_link->tr_list->trans;
-
-		if (tr == NULL) {
-			tr = (fbird_transaction *) emalloc(sizeof(fbird_transaction));
-			tr->handle.ptr = 0;
-			tr->link_cnt = 1;
-			tr->affected_rows = 0;
-			tr->fbt_transaction = NULL;
-			tr->db_link[0] = ib_link;
-			ib_link->tr_list->trans = tr;
-		}
-		if (tr->handle.ptr == 0) {
-			zend_long trans_argl = IBG(default_trans_params);
-			char last_tpb[TPB_MAX_SIZE];
-			unsigned short tpb_len = 0;
-
-			/* Build TPB if non-default parameters */
-			if (trans_argl != PHP_FBIRD_DEFAULT) {
-				zend_long trans_timeout = IBG(default_lock_timeout);
-				_php_fbird_populate_trans(trans_argl, trans_timeout, last_tpb, &tpb_len);
-			}
-
-			/* OO API Only: All connections use fbt_start() */
-			if (ib_link->fbc_connection == NULL) {
-				_php_fbird_module_error("Connection has no OO API handle");
-				return FAILURE;
-			}
-
-			void* attachment = fbc_get_attachment(ib_link->fbc_connection);
-			if (attachment == NULL) {
-				_php_fbird_module_error("Failed to get attachment from OO API connection");
-				return FAILURE;
-			}
-
-			tr->fbt_transaction = fbt_start(
-				IBG(master_instance),
-				attachment,
-				tpb_len,
-				tpb_len > 0 ? (const unsigned char*)last_tpb : NULL,
-				IB_STATUS
-			);
-
-			if (tr->fbt_transaction == NULL) {
-				_php_fbird_error();
-				return FAILURE;
-			}
-
-			/* Store a compatible handle for legacy code paths that may inspect it */
-			tr->handle.ptr = fbt_get_handle(tr->fbt_transaction);
-		}
-		*trans = tr;
-	}
-	return SUCCESS;
-}
-
-static void _php_fbird_trans_end(INTERNAL_FUNCTION_PARAMETERS, int commit)
-{
-	fbird_transaction *trans = NULL;
-	int res_id = 0;
-	ISC_STATUS result;
-	fbird_db_link *ib_link;
-	zval *arg = NULL;
-
-	RESET_ERRMSG;
-
-	if (zend_parse_parameters(ZEND_NUM_ARGS(), "|r", &arg) == FAILURE) {
-		return;
-	}
-
-	if (ZEND_NUM_ARGS() == 0) {
-		ib_link = (fbird_db_link *)zend_fetch_resource2(IBG(default_link), LE_LINK, le_link, le_plink);
-		if (ib_link->tr_list == NULL || ib_link->tr_list->trans == NULL) {
-			/* this link doesn't have a default transaction */
-			_php_fbird_module_error("Default link has no default transaction");
-			RETURN_FALSE;
-		}
-		trans = ib_link->tr_list->trans;
-	} else {
-		/* one id was passed, could be db or trans id */
-		if (Z_RES_P(arg)->type == le_trans) {
-			trans = (fbird_transaction *)zend_fetch_resource_ex(arg, LE_TRANS, le_trans);
-			res_id = Z_RES_P(arg)->handle;
-		} else {
-			ib_link = (fbird_db_link *)zend_fetch_resource2_ex(arg, LE_LINK, le_link, le_plink);
-
-			if (ib_link->tr_list == NULL || ib_link->tr_list->trans == NULL) {
-				/* this link doesn't have a default transaction */
-				_php_fbird_module_error("Link has no default transaction");
-				RETURN_FALSE;
-			}
-			trans = ib_link->tr_list->trans;
-		}
-	}
-
-	/* OO API Only: All transactions use fbt_* functions */
-	if (trans->fbt_transaction == NULL) {
-		_php_fbird_module_error("invalid transaction handle (expecting explicit transaction start) ");
-		RETURN_FALSE;
-	}
-
-	switch (commit) {
-		default: /* == case ROLLBACK: */
-			result = fbt_rollback(trans->fbt_transaction, IB_STATUS);
-			break;
-		case COMMIT:
-			result = fbt_commit(trans->fbt_transaction, IB_STATUS);
-			break;
-		case (ROLLBACK | RETAIN):
-			result = fbt_rollback_retaining(trans->fbt_transaction, IB_STATUS);
-			break;
-		case (COMMIT | RETAIN):
-			result = fbt_commit_retaining(trans->fbt_transaction, IB_STATUS);
-			break;
-	}
-
-	/* Clear handle for non-retained operations BEFORE checking result.
-	 * The fbt_* functions ALWAYS delete the wrapper (even on error),
-	 * so we must clear our pointer to avoid dangling references.
-	 * Fixes: #9, #10 - SIGSEGV due to use-after-free of transaction wrapper
-	 */
-	if ((commit & RETAIN) == 0) {
-		fbt_free(trans->fbt_transaction);
-		trans->fbt_transaction = NULL;
-		trans->handle.ptr = 0;
-	}
-
-	if (result) {
-		_php_fbird_error();
-		RETURN_FALSE;
-	}
-
-	/* Don't try to destroy implicitly opened transaction from list... */
-	if ((commit & RETAIN) == 0 && res_id != 0) {
-		zend_list_delete(Z_RES_P(arg));
-	}
-	RETURN_TRUE;
-}
-
-PHP_FUNCTION(fbird_commit)
-{
-	_php_fbird_trans_end(INTERNAL_FUNCTION_PARAM_PASSTHRU, COMMIT);
-}
-
-PHP_FUNCTION(fbird_rollback)
-{
-	_php_fbird_trans_end(INTERNAL_FUNCTION_PARAM_PASSTHRU, ROLLBACK);
-}
-
-PHP_FUNCTION(fbird_commit_ret)
-{
-	_php_fbird_trans_end(INTERNAL_FUNCTION_PARAM_PASSTHRU, COMMIT | RETAIN);
-}
-
-PHP_FUNCTION(fbird_rollback_ret)
-{
-	_php_fbird_trans_end(INTERNAL_FUNCTION_PARAM_PASSTHRU, ROLLBACK | RETAIN);
-}
-
-static int is_valid_identifier(const char *s, size_t len)
-{
-	size_t i;
-	if (len == 0) return 0;
-
-	if (s[0] == '"') {
-		/* Quoted identifier: must end with quote, and internal quotes must be paired */
-		if (len < 2 || s[len-1] != '"') return 0;
-		for (i = 1; i < len - 1; i++) {
-			if (s[i] == '"') {
-				if (s[i+1] == '"') {
-					i++; /* skip paired quote */
-				} else {
-					return 0; /* unescaped quote in middle */
-				}
-			}
-		}
-		return 1;
-	}
-
-	/* Unquoted identifier: alphanumeric, _, $ */
-	for (i = 0; i < len; i++) {
-		char c = s[i];
-		if (!((c >= 'A' && c <= 'Z') || (c >= 'a' && c <= 'z') ||
-			  (c >= '0' && c <= '9') || c == '_' || c == '$')) {
-			return 0;
-		}
-	}
-	return 1;
-}
 
 PHP_FUNCTION(fbird_gen_id)
 {
@@ -3050,619 +1199,3 @@ PHP_FUNCTION(fbird_reconnect_transaction)
 
 #if FB_API_VER >= 40
 /* IBatch API Functions (Firebird 4.0+ Bulk Operations) */
-PHP_FUNCTION(fbird_batch_create)
-{
-	zval *query_arg, *trans_arg = NULL;
-	fbird_query *ib_query;
-	fbird_transaction *trans = NULL;
-	fbird_batch *ib_batch;
-	void *stmt_ptr;
-	void *batch_wrapper;
-	void *metadata;
-	unsigned msg_length;
-
-	RESET_ERRMSG;
-
-	if (zend_parse_parameters(ZEND_NUM_ARGS(), "r|r!", &query_arg, &trans_arg) == FAILURE) {
-		return;
-	}
-
-	ib_query = (fbird_query *)zend_fetch_resource_ex(query_arg, LE_QUERY, le_query);
-	if (!ib_query) {
-		RETURN_FALSE;
-	}
-
-	if (!ib_query->fbs_statement) {
-		_php_fbird_module_error("Query has no OO API statement handle");
-		RETURN_FALSE;
-	}
-
-	/* Get transaction - either from parameter or from query's default */
-	if (trans_arg != NULL) {
-		trans = (fbird_transaction *)zend_fetch_resource_ex(trans_arg, LE_TRANS, le_trans);
-		if (!trans) {
-			RETURN_FALSE;
-		}
-	} else {
-		trans = ib_query->trans;
-	}
-
-	if (!trans || !trans->fbt_transaction) {
-		_php_fbird_module_error("No valid transaction for batch operation");
-		RETURN_FALSE;
-	}
-
-	/* Get raw IStatement pointer */
-	stmt_ptr = fbs_get_statement(ib_query->fbs_statement);
-	if (!stmt_ptr) {
-		_php_fbird_module_error("Failed to get statement handle");
-		RETURN_FALSE;
-	}
-
-	/* Create batch with default buffer size */
-	batch_wrapper = fbbatch_create(IBG(master_instance), stmt_ptr, 0, IB_STATUS);
-	if (!batch_wrapper) {
-		_php_fbird_error();
-		RETURN_FALSE;
-	}
-
-	/* Get input metadata for building message buffers */
-	metadata = fbbatch_get_metadata(IBG(master_instance), batch_wrapper, IB_STATUS);
-	if (!metadata) {
-		fbbatch_close(IBG(master_instance), batch_wrapper, IB_STATUS);
-		_php_fbird_error();
-		RETURN_FALSE;
-	}
-
-	msg_length = fbm_get_message_length(IBG(master_instance), metadata);
-
-	/* Allocate batch structure */
-	ib_batch = (fbird_batch *)ecalloc(1, sizeof(fbird_batch));
-	ib_batch->fbbatch_wrapper = batch_wrapper;
-	ib_batch->trans = trans;
-	ib_batch->query = ib_query;
-	ib_batch->in_metadata = metadata;
-	ib_batch->in_msg_length = msg_length;
-	ib_batch->in_msg_buffer = emalloc(msg_length);
-	memset(ib_batch->in_msg_buffer, 0, msg_length);
-
-	RETVAL_RES(zend_register_resource(ib_batch, le_batch));
-}
-
-PHP_FUNCTION(fbird_batch_add)
-{
-	zval *batch_arg;
-	zval *args = NULL;
-	int argc = 0;
-	fbird_batch *ib_batch;
-
-	RESET_ERRMSG;
-
-	if (zend_parse_parameters(ZEND_NUM_ARGS(), "r*", &batch_arg, &args, &argc) == FAILURE) {
-		return;
-	}
-
-	ib_batch = (fbird_batch *)zend_fetch_resource_ex(batch_arg, LE_BATCH, le_batch);
-	if (!ib_batch || !ib_batch->fbbatch_wrapper) {
-		php_error_docref(NULL, E_WARNING, "Invalid batch resource");
-		RETURN_FALSE;
-	}
-
-	/* Get master interface for metadata operations */
-	void *master = IBG(master_instance);
-	if (!master) {
-		_php_fbird_module_error("fbird_batch_add() requires Firebird 3.0+ OO API master interface");
-		RETURN_FALSE;
-	}
-
-	/* Validate metadata and buffer are available */
-	if (!ib_batch->in_metadata || !ib_batch->in_msg_buffer || ib_batch->in_msg_length == 0) {
-		_php_fbird_module_error("fbird_batch_add() batch has no input metadata or buffer");
-		RETURN_FALSE;
-	}
-
-	/* Get parameter count from metadata */
-	unsigned param_count = fbm_get_count(master, ib_batch->in_metadata);
-
-	/* Validate argument count matches parameter count */
-	if ((unsigned)argc != param_count) {
-		_php_fbird_module_error("fbird_batch_add() expects %u parameters, %d given", param_count, argc);
-		RETURN_FALSE;
-	}
-
-	/* Clear message buffer before populating */
-	memset(ib_batch->in_msg_buffer, 0, ib_batch->in_msg_length);
-
-	/* Bind each parameter to the message buffer */
-	for (unsigned i = 0; i < param_count; i++) {
-		zval *b_var = &args[i];
-
-		/* Get metadata info for this parameter */
-		unsigned sql_type = fbm_get_type(master, ib_batch->in_metadata, i) & ~1;
-		unsigned data_offset = fbm_get_offset(master, ib_batch->in_metadata, i);
-		unsigned null_offset = fbm_get_null_offset(master, ib_batch->in_metadata, i);
-		unsigned field_length = fbm_get_length(master, ib_batch->in_metadata, i);
-		int sql_scale = fbm_get_scale(master, ib_batch->in_metadata, i);
-
-		/* Get pointers to data and null indicator in message buffer */
-		unsigned char *data_ptr = (unsigned char *)ib_batch->in_msg_buffer + data_offset;
-		short *null_ptr = (short *)((unsigned char *)ib_batch->in_msg_buffer + null_offset);
-
-		/* Handle NULL values */
-		int is_null = 0;
-		switch (Z_TYPE_P(b_var)) {
-			case IS_NULL:
-				is_null = 1;
-				break;
-			case IS_STRING:
-				/* Empty string treated as NULL for numeric/date types */
-				if (Z_STRLEN_P(b_var) == 0) {
-					switch (sql_type) {
-						case SQL_SHORT:
-						case SQL_LONG:
-						case SQL_INT64:
-						case SQL_FLOAT:
-						case SQL_DOUBLE:
-						case SQL_TIMESTAMP:
-						case SQL_TYPE_DATE:
-						case SQL_TYPE_TIME:
-#if FB_API_VER >= 40
-						case SQL_INT128:
-						case SQL_DEC16:
-						case SQL_DEC34:
-						case SQL_TIMESTAMP_TZ:
-						case SQL_TIME_TZ:
-#endif
-							is_null = 1;
-							break;
-						default:
-							break;
-					}
-				}
-				break;
-			default:
-				break;
-		}
-
-		if (is_null) {
-			*null_ptr = -1;
-			continue;
-		}
-
-		/* Not NULL */
-		*null_ptr = 0;
-
-		/* Convert PHP value to Firebird format based on SQL type */
-		switch (sql_type) {
-			case SQL_SHORT: {
-				if (sql_scale < 0) {
-					/* NUMERIC/DECIMAL with scale */
-					double dval = zval_get_double(b_var);
-					double factor = pow(10.0, (double)(-sql_scale));
-					long long scaled = llround(dval * factor);
-					if (scaled < SHRT_MIN || scaled > SHRT_MAX) {
-						_php_fbird_module_error("Parameter %u: scaled value out of range for SHORT", i + 1);
-						RETURN_FALSE;
-					}
-					*(short *)data_ptr = (short)scaled;
-				} else {
-					zend_long lval = zval_get_long(b_var);
-					*(short *)data_ptr = (short)lval;
-				}
-				break;
-			}
-
-			case SQL_LONG: {
-				if (sql_scale < 0) {
-					double dval = zval_get_double(b_var);
-					double factor = pow(10.0, (double)(-sql_scale));
-					long long scaled = llround(dval * factor);
-					if (scaled < INT_MIN || scaled > INT_MAX) {
-						_php_fbird_module_error("Parameter %u: scaled value out of range for LONG", i + 1);
-						RETURN_FALSE;
-					}
-					*(ISC_LONG *)data_ptr = (ISC_LONG)scaled;
-				} else {
-					zend_long lval = zval_get_long(b_var);
-					*(ISC_LONG *)data_ptr = (ISC_LONG)lval;
-				}
-				break;
-			}
-
-			case SQL_INT64: {
-				if (sql_scale < 0) {
-					double dval = zval_get_double(b_var);
-					double factor = pow(10.0, (double)(-sql_scale));
-					*(ISC_INT64 *)data_ptr = (ISC_INT64)llround(dval * factor);
-				} else {
-					zend_long lval = zval_get_long(b_var);
-					*(ISC_INT64 *)data_ptr = (ISC_INT64)lval;
-				}
-				break;
-			}
-
-			case SQL_FLOAT: {
-				double dval = zval_get_double(b_var);
-				*(float *)data_ptr = (float)dval;
-				break;
-			}
-
-			case SQL_DOUBLE: {
-				double dval = zval_get_double(b_var);
-				*(double *)data_ptr = dval;
-				break;
-			}
-
-			case SQL_TEXT: {
-				/* Fixed-length CHAR field */
-				convert_to_string(b_var);
-				size_t str_len = Z_STRLEN_P(b_var);
-				if (str_len > field_length) {
-					str_len = field_length;
-				}
-				memcpy(data_ptr, Z_STRVAL_P(b_var), str_len);
-				/* Pad with spaces for CHAR type */
-				if (str_len < field_length) {
-					memset(data_ptr + str_len, ' ', field_length - str_len);
-				}
-				break;
-			}
-
-			case SQL_VARYING: {
-				/* VARCHAR: 2-byte length prefix + data */
-				convert_to_string(b_var);
-				size_t str_len = Z_STRLEN_P(b_var);
-				if (str_len > field_length) {
-					str_len = field_length;
-				}
-				*(short *)data_ptr = (short)str_len;
-				memcpy(data_ptr + sizeof(short), Z_STRVAL_P(b_var), str_len);
-				break;
-			}
-
-			case SQL_TIMESTAMP:
-			case SQL_TYPE_DATE:
-			case SQL_TYPE_TIME: {
-				if (Z_TYPE_P(b_var) == IS_LONG) {
-					/* Unix timestamp */
-					struct tm t;
-					time_t ts = (time_t)Z_LVAL_P(b_var);
-					struct tm *res = php_gmtime_r(&ts, &t);
-					if (!res) {
-						_php_fbird_module_error("Parameter %u: invalid timestamp value", i + 1);
-						RETURN_FALSE;
-					}
-					switch (sql_type) {
-						case SQL_TIMESTAMP:
-							*(ISC_TIMESTAMP *)data_ptr = fbu_encode_timestamp(master,
-								(unsigned)(t.tm_year + 1900), (unsigned)(t.tm_mon + 1),
-								(unsigned)t.tm_mday, (unsigned)t.tm_hour,
-								(unsigned)t.tm_min, (unsigned)t.tm_sec, 0);
-							break;
-						case SQL_TYPE_DATE:
-							*(ISC_DATE *)data_ptr = fbu_encode_date(master,
-								(unsigned)(t.tm_year + 1900), (unsigned)(t.tm_mon + 1),
-								(unsigned)t.tm_mday);
-							break;
-						case SQL_TYPE_TIME:
-							*(ISC_TIME *)data_ptr = fbu_encode_time(master,
-								(unsigned)t.tm_hour, (unsigned)t.tm_min,
-								(unsigned)t.tm_sec, 0);
-							break;
-					}
-				} else {
-					/* Parse string date/time */
-					convert_to_string(b_var);
-					fbird_datetime_components dt;
-					int parsed = 0;
-					switch (sql_type) {
-						case SQL_TYPE_DATE:
-							parsed = fbird_parse_date(Z_STRVAL_P(b_var), &dt);
-							if (parsed) {
-								*(ISC_DATE *)data_ptr = fbu_encode_date(master, dt.year, dt.month, dt.day);
-							}
-							break;
-						case SQL_TYPE_TIME:
-							parsed = fbird_parse_time(Z_STRVAL_P(b_var), &dt);
-							if (parsed) {
-								*(ISC_TIME *)data_ptr = fbu_encode_time(master, dt.hours, dt.minutes, dt.seconds, dt.fractions);
-							}
-							break;
-						default: /* SQL_TIMESTAMP */
-							parsed = fbird_parse_timestamp(Z_STRVAL_P(b_var), &dt);
-							if (parsed) {
-								*(ISC_TIMESTAMP *)data_ptr = fbu_encode_timestamp(master,
-									dt.year, dt.month, dt.day, dt.hours, dt.minutes, dt.seconds, dt.fractions);
-							}
-							break;
-					}
-					if (!parsed) {
-						_php_fbird_module_error("Parameter %u: invalid date/time string '%s'", i + 1, Z_STRVAL_P(b_var));
-						RETURN_FALSE;
-					}
-				}
-				break;
-			}
-
-#if FB_API_VER >= 40
-			case SQL_TIMESTAMP_TZ:
-			case SQL_TIME_TZ: {
-				fbird_datetime_components dt;
-				fbird_datetime_init(&dt);
-				strncpy(dt.timezone, "GMT", sizeof(dt.timezone) - 1);
-
-				if (Z_TYPE_P(b_var) == IS_LONG) {
-					struct tm t;
-					time_t ts = (time_t)Z_LVAL_P(b_var);
-					struct tm *res = php_gmtime_r(&ts, &t);
-					if (!res) {
-						_php_fbird_module_error("Parameter %u: invalid timestamp value", i + 1);
-						RETURN_FALSE;
-					}
-					dt.year = (unsigned)(t.tm_year + 1900);
-					dt.month = (unsigned)(t.tm_mon + 1);
-					dt.day = (unsigned)t.tm_mday;
-					dt.hours = (unsigned)t.tm_hour;
-					dt.minutes = (unsigned)t.tm_min;
-					dt.seconds = (unsigned)t.tm_sec;
-				} else {
-					convert_to_string(b_var);
-					int parsed = (sql_type == SQL_TIME_TZ)
-						? fbird_parse_time(Z_STRVAL_P(b_var), &dt)
-						: fbird_parse_timestamp(Z_STRVAL_P(b_var), &dt);
-					if (!parsed) {
-						_php_fbird_module_error("Parameter %u: invalid date/time string", i + 1);
-						RETURN_FALSE;
-					}
-					if (!dt.has_timezone) {
-						strncpy(dt.timezone, "GMT", sizeof(dt.timezone) - 1);
-					}
-				}
-
-				if (sql_type == SQL_TIME_TZ) {
-					if (fbu_encode_time_tz(master, (ISC_TIME_TZ *)data_ptr,
-							dt.hours, dt.minutes, dt.seconds, dt.fractions, dt.timezone) != 0) {
-						_php_fbird_module_error("Parameter %u: failed to encode TIME WITH TIME ZONE", i + 1);
-						RETURN_FALSE;
-					}
-				} else {
-					if (fbu_encode_timestamp_tz(master, (ISC_TIMESTAMP_TZ *)data_ptr,
-							dt.year, dt.month, dt.day, dt.hours, dt.minutes, dt.seconds,
-							dt.fractions, dt.timezone) != 0) {
-						_php_fbird_module_error("Parameter %u: failed to encode TIMESTAMP WITH TIME ZONE", i + 1);
-						RETURN_FALSE;
-					}
-				}
-				break;
-			}
-#endif
-
-#ifdef SQL_BOOLEAN
-			case SQL_BOOLEAN: {
-				FB_BOOLEAN bval;
-				switch (Z_TYPE_P(b_var)) {
-					case IS_TRUE:
-						bval = FB_TRUE;
-						break;
-					case IS_FALSE:
-						bval = FB_FALSE;
-						break;
-					case IS_LONG:
-					case IS_DOUBLE:
-						bval = zend_is_true(b_var) ? FB_TRUE : FB_FALSE;
-						break;
-					case IS_STRING:
-						if (Z_STRLEN_P(b_var) == 0) {
-							bval = FB_FALSE;
-						} else if (!zend_binary_strncasecmp(Z_STRVAL_P(b_var), Z_STRLEN_P(b_var), "true", 4, 4)) {
-							bval = FB_TRUE;
-						} else if (!zend_binary_strncasecmp(Z_STRVAL_P(b_var), Z_STRLEN_P(b_var), "false", 5, 5)) {
-							bval = FB_FALSE;
-						} else {
-							zend_long lval;
-							double dval;
-							switch (is_numeric_string(Z_STRVAL_P(b_var), Z_STRLEN_P(b_var), &lval, &dval, 0)) {
-								case IS_LONG:
-									bval = (lval != 0) ? FB_TRUE : FB_FALSE;
-									break;
-								case IS_DOUBLE:
-									bval = (dval != 0) ? FB_TRUE : FB_FALSE;
-									break;
-								default:
-									_php_fbird_module_error("Parameter %u: cannot convert string to boolean", i + 1);
-									RETURN_FALSE;
-							}
-						}
-						break;
-					default:
-						bval = zend_is_true(b_var) ? FB_TRUE : FB_FALSE;
-						break;
-				}
-				*(FB_BOOLEAN *)data_ptr = bval;
-				break;
-			}
-#endif
-
-			case SQL_BLOB: {
-				/* BLOB ID as hex string "HHHHHHHH:LLLL" (13 characters)
-				 * Format: 8 hex digits (high 32-bit), colon, 4 hex digits (low 16-bit)
-				 * Example: "74292B00:7FFC"
-				 */
-				convert_to_string(b_var);
-
-				if (Z_STRLEN_P(b_var) == BLOB_ID_LEN &&
-					_php_fbird_string_to_quad(Z_STRVAL_P(b_var), (ISC_QUAD *)data_ptr)) {
-					/* Valid BLOB ID parsed and written to message buffer */
-					break;
-				}
-
-				/* Invalid BLOB ID format */
-				_php_fbird_module_error("Parameter %u: BLOB must be passed as blob ID (use fbird_batch_add_blob() or fbird_blob_create())", i + 1);
-				RETURN_FALSE;
-			}
-
-			default:
-				_php_fbird_module_error("Parameter %u: unsupported SQL type %u for batch binding", i + 1, sql_type);
-				RETURN_FALSE;
-		}
-	}
-
-	/* Add the populated message buffer to the batch */
-	if (fbbatch_add(master, ib_batch->fbbatch_wrapper, 1, ib_batch->in_msg_buffer, IB_STATUS) != 1) {
-		_php_fbird_error();
-		RETURN_FALSE;
-	}
-
-	RETURN_TRUE;
-}
-
-PHP_FUNCTION(fbird_batch_execute)
-{
-	zval *batch_arg;
-	fbird_batch *ib_batch;
-	void *trans_ptr;
-	unsigned total_processed = 0;
-	unsigned error_count = 0;
-
-	RESET_ERRMSG;
-
-	if (zend_parse_parameters(ZEND_NUM_ARGS(), "r", &batch_arg) == FAILURE) {
-		return;
-	}
-
-	ib_batch = (fbird_batch *)zend_fetch_resource_ex(batch_arg, LE_BATCH, le_batch);
-	if (!ib_batch || !ib_batch->fbbatch_wrapper) {
-		php_error_docref(NULL, E_WARNING, "Invalid batch resource");
-		RETURN_FALSE;
-	}
-
-	if (!ib_batch->trans || !ib_batch->trans->fbt_transaction) {
-		_php_fbird_module_error("Batch has no valid transaction");
-		RETURN_FALSE;
-	}
-
-	trans_ptr = fbt_get_handle(ib_batch->trans->fbt_transaction);
-	if (!trans_ptr) {
-		_php_fbird_module_error("Failed to get transaction handle");
-		RETURN_FALSE;
-	}
-
-	if (!fbbatch_execute(IBG(master_instance), ib_batch->fbbatch_wrapper, trans_ptr,
-			&total_processed, &error_count, IB_STATUS)) {
-		_php_fbird_error();
-		RETURN_FALSE;
-	}
-
-	/* Close the batch after execution */
-	fbbatch_close(IBG(master_instance), ib_batch->fbbatch_wrapper, IB_STATUS);
-	ib_batch->fbbatch_wrapper = NULL;
-
-	/* Calculate success_count from total_processed - error_count */
-	unsigned success_count = (total_processed >= error_count) ? (total_processed - error_count) : 0;
-
-	array_init(return_value);
-	add_assoc_long(return_value, "total_processed", total_processed);
-	add_assoc_long(return_value, "success_count", success_count);
-	add_assoc_long(return_value, "error_count", error_count);
-}
-
-PHP_FUNCTION(fbird_batch_cancel)
-{
-	zval *batch_arg;
-	fbird_batch *ib_batch;
-
-	RESET_ERRMSG;
-
-	if (zend_parse_parameters(ZEND_NUM_ARGS(), "r", &batch_arg) == FAILURE) {
-		return;
-	}
-
-	ib_batch = (fbird_batch *)zend_fetch_resource_ex(batch_arg, LE_BATCH, le_batch);
-	if (!ib_batch || !ib_batch->fbbatch_wrapper) {
-		php_error_docref(NULL, E_WARNING, "Invalid batch resource");
-		RETURN_FALSE;
-	}
-
-	if (!fbbatch_cancel(IBG(master_instance), ib_batch->fbbatch_wrapper, IB_STATUS)) {
-		_php_fbird_error();
-		RETURN_FALSE;
-	}
-
-	fbbatch_close(IBG(master_instance), ib_batch->fbbatch_wrapper, IB_STATUS);
-	ib_batch->fbbatch_wrapper = NULL;
-
-	RETURN_TRUE;
-}
-
-PHP_FUNCTION(fbird_batch_add_blob)
-{
-	zval *batch_arg;
-	char *data;
-	size_t data_len;
-	zend_long blob_type = 0;
-	fbird_batch *ib_batch;
-	ISC_QUAD blob_id;
-
-	RESET_ERRMSG;
-
-	if (zend_parse_parameters(ZEND_NUM_ARGS(), "rs|l", &batch_arg, &data, &data_len, &blob_type) == FAILURE) {
-		return;
-	}
-
-	ib_batch = (fbird_batch *)zend_fetch_resource_ex(batch_arg, LE_BATCH, le_batch);
-	if (!ib_batch || !ib_batch->fbbatch_wrapper) {
-		php_error_docref(NULL, E_WARNING, "Invalid batch resource");
-		RETURN_FALSE;
-	}
-
-	/* Call C++ wrapper to add BLOB to batch */
-	if (fbbatch_add_blob(IBG(master_instance), ib_batch->fbbatch_wrapper,
-			(unsigned)data_len, data, &blob_id, 0, NULL, IB_STATUS) == 0) {
-		_php_fbird_error();
-		RETURN_FALSE;
-	}
-
-	/* Convert BLOB ID to hex string for PHP */
-	RETURN_NEW_STR(_php_fbird_quad_to_string(blob_id));
-}
-
-PHP_FUNCTION(fbird_batch_register_blob)
-{
-	zval *batch_arg;
-	char *blob_id_str;
-	size_t blob_id_len;
-	fbird_batch *ib_batch;
-	ISC_QUAD existing_blob, batch_blob_id;
-
-	RESET_ERRMSG;
-
-	if (zend_parse_parameters(ZEND_NUM_ARGS(), "rs", &batch_arg, &blob_id_str, &blob_id_len) == FAILURE) {
-		return;
-	}
-
-	ib_batch = (fbird_batch *)zend_fetch_resource_ex(batch_arg, LE_BATCH, le_batch);
-	if (!ib_batch || !ib_batch->fbbatch_wrapper) {
-		php_error_docref(NULL, E_WARNING, "Invalid batch resource");
-		RETURN_FALSE;
-	}
-
-	/* Validate and convert BLOB ID string to ISC_QUAD */
-	if (blob_id_len != BLOB_ID_LEN || !_php_fbird_string_to_quad(blob_id_str, &existing_blob)) {
-		php_error_docref(NULL, E_WARNING, "Invalid BLOB ID format (expected %d character hex string)", BLOB_ID_LEN);
-		RETURN_FALSE;
-	}
-
-	/* Call C++ wrapper to register BLOB in batch */
-	if (fbbatch_register_blob(IBG(master_instance), ib_batch->fbbatch_wrapper,
-			&existing_blob, &batch_blob_id, IB_STATUS) == 0) {
-		_php_fbird_error();
-		RETURN_FALSE;
-	}
-
-	/* Convert batch BLOB ID to hex string for PHP */
-	RETURN_NEW_STR(_php_fbird_quad_to_string(batch_blob_id));
-}
-#endif /* FB_API_VER >= 40 */
-
-
-#endif /* HAVE_FIREBIRD */
